refactor SegmentFile and introduce SegmentFileWriter
- Separate constructors and methods for writing from SegmentFile to SegmentFileWriter
